+++ packages/carbon-graphs/src/js/helpers/axis.js
@@ -897,19 +897,17 @@ const createAxes = (axis, scale, config, canvasSVG) => {
     .attr('id', 'x')
     .selectAll('text')
     .style('text-anchor', () => {
-      const validRotations = Object.values(utils.TickLabelRotations);
       if (
-        utils.isDefined(config.axis.x.ticks.TickLabelRotations)
-        && validRotations.includes(config.axis.x.ticks.TickLabelRotations)
+        utils.isDefined(config.axis.x.ticks.tickLabelsRotation)
+        && utils.validTickLabelRotations.has(config.axis.x.ticks.tickLabelsRotation)
       ) {
-        const rotation = config.axis.x.ticks.TickLabelRotations;
-        console.warn(`tickLabelsRotation Using ${rotation} rotation for x-axis labels to avoid overlapping.`);
-        return rotation === utils.TickLabelRotations.NEGATIVE_45 ? 'end' : 'middle';
+        const rotation = config.axis.x.ticks.tickLabelsRotation;
+        return rotation === -45 ? 'end' : 'middle';
       }
-      console.warn('Warning: Invalid tickLabelsRotation angle. Using default rotation (0 or -45) for x-axis labels.');
+      console.warn('Warning: Invalid tickLabelsRotation angle. Valid values are (0 or -45) for x-axis labels.');
       return 'middle';
     })
-    .attr('transform', () => `rotate(${config.axis.x.ticks.TickLabelRotations})`);
+    .attr('transform', () => `rotate(${config.axis.x.ticks.tickLabelsRotation})`);
   canvasSVG
     .append('g')
     .classed(styles.axis, true)
@@ -1137,13 +1135,12 @@ const getXAxisHeight = (config) => {
   const dummy = d3.select('body').append('div');
   const svg = dummy.append('svg');
   const group = svg.append('g').call(axis);
-  if (config.axis.x.ticks && config.axis.x.ticks.TickLabelRotations === -45) {
+  if (config.axis.x.ticks && config.axis.x.ticks.tickLabelsRotation === -45) {
     // Add extra padding for rotated tick labels
-    const extraPadding = 5;
-    group.selectAll('.tick text').attr('transform', `rotate(${config.axis.x.ticks.TickLabelRotations})`);
-    const rotatedHeight = group.node().getBoundingClientRect().height;
+    group.selectAll('.tick text').attr('transform', `rotate(${config.axis.x.ticks.tickLabelsRotation})`);
+    const rotatedTickLabelsHeight = group.node().getBoundingClientRect().height;
     dummy.remove();
-    return rotatedHeight + extraPadding;
+    return rotatedTickLabelsHeight + constants.DEFAULT_OVERLAPPING_PADDING;
   }
   const { height } = group.node().getBoundingClientRect();
   dummy.remove();
@@ -1228,14 +1225,14 @@ const getAxisLabelWidth = (label, axis, config) => {
  * @param {string} label - Label text
  * @returns {number} label height
  */
-const getAxisLabelHeight = (label, TickLabelRotations) => {
+const getAxisLabelHeight = (label, tickLabelsRotation) => {
   const dummy = d3.select('body').append('div');
   const svg = dummy.append('svg');
   const grouper = svg.append('g');
 
-  if (TickLabelRotations && TickLabelRotations === -45) {
+  if (tickLabelsRotation === -45) {
     // Adding extra padding for rotated labels
-    grouper.attr('transform', `rotate(${TickLabelRotations})`);
+    grouper.attr('transform', `rotate(${tickLabelsRotation})`);
   }
 
   buildAxisLabel(grouper, label);
@@ -1322,8 +1319,8 @@ const calculateAxesLabelSize = (config) => {
   config.axisLabelWidths.y2 = 0;
   if (config.showLabel) {
     if (config.axis.x.label) {
-      const TickLabelRotations = config.axis.x.ticks && config.axis.x.ticks.TickLabelRotations;
-      config.axisLabelHeights.x = getAxisLabelHeight(config.axis.x.label, TickLabelRotations);
+      const tickLabelsRotation = config.axis.x.ticks && config.axis.x.ticks.tickLabelsRotation;
+      config.axisLabelHeights.x = getAxisLabelHeight(config.axis.x.label, tickLabelsRotation);
     }
     if (config.axis.y.label) {
       config.axisLabelWidths.y = constants.DEFAULT_CHARACTER_SVG_ELEMENT_WIDTH;