+++ packages/carbon-graphs/CHANGELOG.md
@@ -1,8 +1,9 @@
 # ChangeLog
 
 ## Unreleased
-* Fixed
-  * Added prop `rotateAngle` to allow rotation of x-axis tick labels to prevent overlap.
+
+* Added
+  * Added prop `tickLabelsRotation` to allow rotation of x-axis tick labels to prevent overlap.
 
 ## 2.23.3 - (September 25, 2023)